Second Activity Correctly Refactored for Naming Purposes
