+++ app/src/main/java/com/example/orientex_v7/Logged_In_View.kt
@@ -26,7 +26,7 @@ class Logged_In_View : AppCompatActivity() {
         // menu should be considered as top level destinations.
         val appBarConfiguration = AppBarConfiguration(
             setOf(
-                R.id.navigation_home, R.id.navigation_dashboard, R.id.navigation_notifications
+                R.id.navigation_home, R.id.navigation_quests, R.id.navigation_profile
             )
         )
         setupActionBarWithNavController(navController, appBarConfiguration)
