+++ app/src/main/java/com/example/orientex_v7/ui/profile/ProfileViewModel.kt
@@ -1,10 +1,10 @@
-package com.example.orientex_v7.ui.notifications
+package com.example.orientex_v7.ui.profile
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class NotificationsViewModel : ViewModel() {
+class ProfileViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
         value = "This is notifications Fragment"
