+++ app/src/main/java/com/example/orientex_v7/ui/quests/QuestsFragment.kt
@@ -1,4 +1,4 @@
-package com.example.orientex_v7.ui.dashboard
+package com.example.orientex_v7.ui.quests
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,11 +7,11 @@ import android.view.ViewGroup
 import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
-import com.example.orientex_v7.databinding.FragmentDashboardBinding
+import com.example.orientex_v7.databinding.FragmentQuestsBinding
 
-class DashboardFragment : Fragment() {
+class QuestsFragment : Fragment() {
 
-    private var _binding: FragmentDashboardBinding? = null
+    private var _binding: FragmentQuestsBinding? = null
 
     // This property is only valid between onCreateView and
     // onDestroyView.
@@ -23,12 +23,12 @@ class DashboardFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View {
         val dashboardViewModel =
-            ViewModelProvider(this).get(DashboardViewModel::class.java)
+            ViewModelProvider(this).get(QuestsViewModel::class.java)
 
-        _binding = FragmentDashboardBinding.inflate(inflater, container, false)
+        _binding = FragmentQuestsBinding.inflate(inflater, container, false)
         val root: View = binding.root
 
-        val textView: TextView = binding.textDashboard
+        val textView: TextView = binding.textQuests
         dashboardViewModel.text.observe(viewLifecycleOwner) {
             textView.text = it
         }
