+++ app/src/main/java/com/example/orientex_v7/ui/quests/QuestsViewModel.kt
@@ -1,10 +1,10 @@
-package com.example.orientex_v7.ui.dashboard
+package com.example.orientex_v7.ui.quests
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class DashboardViewModel : ViewModel() {
+class QuestsViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
         value = "This is dashboard Fragment"
