+++ app/src/main/res/layout/fragment_quests.xml
@@ -4,10 +4,10 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.dashboard.DashboardFragment">
+    tools:context=".ui.quests.QuestsFragment">
 
     <TextView
-        android:id="@+id/text_dashboard"
+        android:id="@+id/text_quests"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginStart="8dp"
