+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -12,14 +12,14 @@
         tools:layout="@layout/fragment_home" />
 
     <fragment
-        android:id="@+id/navigation_dashboard"
-        android:name="com.example.orientex_v7.ui.dashboard.DashboardFragment"
-        android:label="@string/title_dashboard"
-        tools:layout="@layout/fragment_dashboard" />
+        android:id="@+id/navigation_quests"
+        android:name="com.example.orientex_v7.ui.quests.QuestsFragment"
+        android:label="@string/title_quests"
+        tools:layout="@layout/fragment_quests" />
 
     <fragment
-        android:id="@+id/navigation_notifications"
-        android:name="com.example.orientex_v7.ui.notifications.NotificationsFragment"
-        android:label="@string/title_notifications"
-        tools:layout="@layout/fragment_notifications" />
+        android:id="@+id/navigation_profile"
+        android:name="com.example.orientex_v7.ui.profile.ProfileFragment"
+        android:label="@string/title_profile"
+        tools:layout="@layout/fragment_profile" />
 </navigation>
